+Improve the `sr-buttons-maxtaglen' function.
It was both inefficient (consing all over the place, using a cl-seq
function...) _and_ hard to read. In a decent Lisp dialect it would
probably be best to reduce directly on `sr-buttons-list', but such
consideration might not apply to Elisp with `max' and `mapcar' being
both C subrs. In any case, we thus got rid of the single runtime cl
dependency.