+Fix the modeline extension desktop support.
When the Sunrise pane buffer is being restored, the other one might
still not be (the naïve use of 0.1s timer is no solution (besides
being totally misguided, as the desktop functionality tried to call
the minor mode function, and thus `sr-modeline-toggle', right away
anyway)) and trying to use the usual pane/window switching functions
is going to result in an error. Meanwhile, `sr-modeline-refresh' does
all that is needed for a single pane.