+Major refactoring of PopViewer extension.
The new `sr-popviewer-select-viewer-action' variable allows to use a
custom function to perform the selection of the viewer and to manage
it in a generic way.

Added new sr-popviewer-style value: single-dedicated-frame.

Redefined semantics of sunrise-cd when in popviewer mode.

New version realizes Tom's usage pattern by simply setting variable
`sr-popviewer-select-viewer-action' to (lambda () (other-window 1)).