+New implementation of history navigation.
The original implementation of history navigation worked quite well most of the
time, but not always. The ring structure caused some of the directories recently
visited to get "lost" under certain conditions. The new implementation is stack
based and strives to maintain the most recently visited entries as near as
possible to the currently displayed entry.