+Improved active/passive path graphical highlighting.
There is no need to generate overlays to highlight the paths in both the active
and the passive pane. We can instead use font lock to first display the path in
both panes with the passive directory face, and then change the appearance of
the path in the active pane using a dynamically built overlay.

Additionally, some times there is a need to use more than one face to display
the active path, e.g. when using simultaneously wdired and the mirror extension
to rename files inside a compressed archive, SC needs to remember both faces and
restore them in order: default face -> mirror face -> wdired face when "going
in", and the other way around when the whole operation is completed. The current
implementation replaces the previous `sr-current-path-face' plain variable with
a list `sr-current-path-faces' to achieve this effect.