Retire the sunrise-checkpoint-old extension
From the comment in sunrise-checkpoint-old.el:

"Beginning with version 4 of the Sunrise Commander, checkpoints were
redefined to be a special form of bookmarks. Unfortunately, creating
bookmarks with custom handlers isn't supported in the version of
bookmarks.el distributed with Emacs 22, so if you use Sunrise
checkpoints and you don't want to update your bookmarks.el, just add
this extension to your .emacs.el to get back the original
functionality."

So it's a compatibility shim for Emacs 22. Since we now require Emacs
24, it has served its time and we can safely remove it.