Remove (defvar sunrise-buttons-command-adapter) from sunrise-tree
refs https://github.com/sunrise-commander/sunrise-commander/issues/90

Normally you should require `sunrise-buttons` where this variable is defined.
The sunrise extension feature file registers itself as sunrise as a
side effect of loading, so we can't do that.
As a better way to avoid this and suppress compiler warnings, check
the existence of variables with boundp.

I think this is a dependency in the wrong direction, so in the end
this is resolved by other means.